refactor(server): consolidate GoalReactor into one canonical module
Per the Effect service conventions: single module with imports, the
Context.Service tag with its interface inline, make, and a canonical
`layer` export; consumers use the namespace import and refer to the
shape as GoalReactor["Service"].
